+++ test/tzdata_test.exs
@@ -1,3 +1,4 @@
 defmodule TzdataTest do
   use ExUnit.Case
+  doctest Tzdata
 end